refactor(codex): split loader responsibilities
Move Codex aggregation, parsing, path discovery, report rendering, speed detection, and serde types into focused modules.

The public adapter entry points and JSON/table report construction stay the same, which keeps output compatibility while making future parser changes easier to review.
